requirements.txt
backports.shutil-get-terminal-size==1.0.0
certifi==2017.7.27.1
chardet==3.0.4
decorator==4.1.2
Django==1.11.3
django-fcm==0.1.1
django-sites==0.9
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
enum34==1.1.6
factory-boy==2.9.2
Faker==0.7.18
fcm-django==0.2.11
idna==2.5
ipaddress==1.0.18
ipython==5.4.1
ipython-genutils==0.2.0
olefile==0.44
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.0
pkg-resources==0.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyfcm==1.3.1
Pygments==2.2.0
PyJWT==1.5.2
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.3
requests-toolbelt==0.8.0
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
whitenoise==3.3.0
